+++ back-end/virtualStore/src/main/resources/db/migration/V03__create_table_tb_acesso.sql
@@ -6,7 +6,7 @@ CREATE SEQUENCE seq_tb_acesso
     CACHE 1;
 
 CREATE TABLE tb_acesso(
-    id BIGINT DEFAULT NEXTVAL('seq_tb_acesso'),
+    id INTEGER DEFAULT NEXTVAL('seq_tb_acesso'),
     descricao VARCHAR(255) NOT NULL,
     CONSTRAINT tb_acesso_pkey PRIMARY KEY (id)
 );
