+++ back-end/virtualStore/src/main/resources/db/migration/V10__create_table_tb_forma_pagamento.sql
@@ -6,7 +6,7 @@ CREATE SEQUENCE seq_tb_forma_pagamento
     CACHE 1;
 
 CREATE TABLE tb_forma_pagamento (
-    id BIGINT DEFAULT NEXTVAL('seq_tb_forma_pagamento'),
+    id INTEGER DEFAULT NEXTVAL('seq_tb_forma_pagamento'),
     descricao VARCHAR(255) NOT NULL,
     CONSTRAINT tb_forma_pagamento_pkey PRIMARY KEY (id)
 );
