+++ back-end/virtualStore/src/main/resources/db/migration/V12__create_table_tb_cupom_desconto.sql
@@ -6,7 +6,7 @@ CREATE SEQUENCE seq_tb_cupom_desconto
     CACHE 1;
 
 CREATE TABLE tb_cupom_desconto (
-    id BIGINT DEFAULT NEXTVAL('seq_tb_cupom_desconto'),
+    id INTEGER DEFAULT NEXTVAL('seq_tb_cupom_desconto'),
     codigo_desconto VARCHAR(255) NOT NULL,
     data_validade_cupom DATE NOT NULL,
     valor_porcentual_desconto NUMERIC(19,2),
